Оптимизации, исправлен баг с загрузкой секторов

+++ World/Core/MapGen/PerlinNoise.cpp
@@ -15,7 +15,7 @@ static void normalize2(float v[2])
 {
   float s;
 
-  s = sqrt(v[0] * v[0] + v[1] * v[1]);
+  s = static_cast<float>(sqrt(v[0] * v[0] + v[1] * v[1]));
   v[0] = v[0] / s;
   v[1] = v[1] / s;
 }
@@ -24,7 +24,7 @@ static void normalize3(float v[3])
 {
   float s;
 
-  s = sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]);
+  s = static_cast<float>(sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]));
   v[0] = v[0] / s;
   v[1] = v[1] / s;
   v[2] = v[2] / s;
